Transaction.Category refactor
